requirements.txt
wheel >= 0.37.1
autopep8 >= 1.6.0
Flask >= 2.0.2
Flask-Migrate >= 3.1.0
Flask-RESTful >= 0.3.9
Flask-SQLAlchemy >= 2.5.1
marshmallow >= 3.14.1
psycopg2-binary >= 2.9.2
pycodestyle >= 2.8.0
python-dotenv >= 0.19.2
SQLAlchemy >= 1.4.28
toml >= 0.10.2
marshmallow-enum >= 1.5.1
Flask-Mail >= 0.9.1
pyjwt >= 2.3.0
flask-bootstrap >= 3.3
PILLOW >= 9.0.1
ldap3 >= 2.9.1
flask-cors >= 3.0.10
pytest >= 7.2
coverage >= 7.0
Flask-CAS >= 1.0.2
Flask-JWT-Extended >= 4.4.4
google-api-python-client >= 2.86.0
click >= 8.1.3
celery[redis] >= 5.2.7
Flask-Limiter >= 3.3.1
